Switch to filtered informers, move controller responsibilities into NewController, add selector labels
