+++ manifests/example/golden-pvc.yaml
@@ -2,6 +2,8 @@ apiVersion: v1
 kind: PersistentVolumeClaim
 metadata:
   name: "golden-pvc"
+  labels:
+    app: "containerized-data-importer"
   annotations:
     kubevirt.io/storage.import.endpoint: ""   # Required.  Format: (http||s3)://www.myUrl.com/path/of/data
     kubevirt.io/storage.import.secretName: "" # Optional.  The name of the secret containing credentials for the data source
